etnewsai/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
httpx==0.27.0
feedparser==6.0.11
beautifulsoup4==4.12.3
requests==2.31.0
chromadb==0.5.23
sentence-transformers==2.7.0
aiosqlite==0.20.0
python-dotenv==1.0.1
apscheduler==3.10.4
pydantic==2.7.1
numpy<2.0.0
